Add a "selectable" option.
"selectable" permits the developer to indicate which range of values may
be selected by end users. When supplied, this allows the "range" option
to indicate the total range of values represented by the slider, and
then limits the user to a subset of those values.

For example:

  $(...).quinn({ range: [0, 10], selectable: [2, 8] });

This creates a slider where the far left value is 0 and the far right is
10, but the user will not be able to move the slider to a value lower
than 2, or higher than 8.