Quinn elements now fill the entire wrapper.
In the past, when you provided a wrapper element to Quinn, it would
reduce the size of the slider "bar" so that both the bar and the handle
fit within the wrapper. As such, you had to set dimensions on the
wrapper sufficiently large to fit both.

Now, the wrapper element should be sized to fit the bar *only*. For
example, the default Quinn theme wrapper element is now expected to be
only 5px high, instead of the previous 22. This means that the vertical
alignment of Quinn will likely look a little out of place on your site.
To adjust for this, add a suitable amount of "margin-top" to the
element.