Split Quinn into separate Quinn, Model and Renderer classes.
This is a major refactoring which splits Quinn into three separate
components which communicate using events. The main $.Quinn is still the
main way to interact with the slider (including setting values), but the
new Model will ensure that the values set are valid, while Renderer
handles the HTML output.