Remove range, selectable in favour of min, max, and drawTo.
This change brings Quinn more in line with the HTML5 range element, and
the jQuery UI slider by requiring that minimum and maximum values are
supplied as "min" and "max" attributes (instead of range: [min, max]).

Previously, selectable allowed you to reduce the range of values which a
user could select. This has been removed in and you should now use the
"drawTo" option to extent the width of the slider. For example, instead
of:

  $('.slider').quinn({
    range: [0, 100], selectable: [ 25, 75 ]
  });

... you would now use:

  $('.slider').quinn({
    min: 25, max: 75, drawTo: { left: 0, right: 100 }
  });