Improve positioning and movement of handles
Instead of trying to work out by how much the handle should overhang
("dangle" over) the edges of the bar, this is now determined by the
left- and right-margin on the handle element. Negative margins will
result in the handle overhanging the edges of the bar.

This has the effect of making rendering of handles more robust when the
width of the quinn element is set directly (instead of being constrained
by a parent element).